Rename package to boltz
