+++ src/boltz/models/simple_cnn.py
