+++ src/boltz/schemas/miner_slice.py
