+++ src/boltz/training/miner.py
@@ -1,8 +1,8 @@
 import torch
-from boltzmann_experimentation.utils.logger import general_logger
+from boltz.utils.logger import general_logger
 
-from boltzmann_experimentation.training.model import Model
-from boltzmann_experimentation.schemas import MinerSlice
+from boltz.training.model import Model
+from boltz.schemas import MinerSlice
 
 
 class Miner:
